improved product deletion logic by adding null check before deletion to return appropriate http status code

+++ src/main/java/org/tarun/ecombackend/controlller/ProductController.java
@@ -76,6 +76,10 @@ public class ProductController {
 
     @DeleteMapping("/product/{productId}")
     public ResponseEntity<String> deleteProduct(@PathVariable int productId){
+        Product product = productService.getProduct(productId);
+        if(product.getId()==-1)
+            return new ResponseEntity<>("Deletion failed. Product not found", HttpStatus.NOT_FOUND);
+
         productService.deleteProduct(productId);
         return new ResponseEntity<>("Success", HttpStatus.OK);
     }
